profile calculations 3 + float error solved

+++ src/main/java/com/test/Kstream/repositories/BankTransactionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface BankTransactionRepository extends JpaRepository<BankTransactionEntity, Long> {
 
 
-     int countBankTransactionEntityByTransactionBankCardNumber(String bankCardNumber);
+     float countBankTransactionEntityByTransactionBankCardNumber(String bankCardNumber);
 
 }
